增加ConnectionInterface接口
