+++ src/db/BaseQuery.php
@@ -13,7 +13,6 @@ declare (strict_types = 1);
 namespace think\db;
 
 use think\Collection;
-use think\db\exception\BindParamException;
 use think\db\exception\DataNotFoundException;
 use think\db\exception\DbException as Exception;
 use think\db\exception\ModelNotFoundException;
@@ -216,44 +215,6 @@ class BaseQuery
     }
 
     /**
-     * 执行查询 返回数据集
-     * @access public
-     * @param string $sql  sql指令
-     * @param array  $bind 参数绑定
-     * @return array
-     * @throws BindParamException
-     * @throws PDOException
-     */
-    public function query(string $sql, array $bind = []): array
-    {
-        return $this->connection->query($this, $sql, $bind);
-    }
-
-    /**
-     * 执行语句
-     * @access public
-     * @param string $sql  sql指令
-     * @param array  $bind 参数绑定
-     * @return int
-     * @throws BindParamException
-     * @throws PDOException
-     */
-    public function execute(string $sql, array $bind = []): int
-    {
-        return $this->connection->execute($this, $sql, $bind, true);
-    }
-
-    /**
-     * 获取返回或者影响的记录数
-     * @access public
-     * @return integer
-     */
-    public function getNumRows(): int
-    {
-        return $this->connection->getNumRows();
-    }
-
-    /**
      * 获取最近一次查询的sql语句
      * @access public
      * @return string
@@ -272,18 +233,6 @@ class BaseQuery
     public function getLastInsID(string $sequence = null)
     {
         return $this->connection->getLastInsID($this, $sequence);
-    }
-
-    /**
-     * 批处理执行SQL语句
-     * 批处理的指令都认为是execute操作
-     * @access public
-     * @param array $sql SQL批处理指令
-     * @return bool
-     */
-    public function batchQuery(array $sql = []): bool
-    {
-        return $this->connection->batchQuery($this, $sql);
     }
 
     /**
@@ -814,30 +763,6 @@ class BaseQuery
     }
 
     /**
-     * USING支持 用于多表删除
-     * @access public
-     * @param mixed $using USING
-     * @return $this
-     */
-    public function using($using)
-    {
-        $this->options['using'] = $using;
-        return $this;
-    }
-
-    /**
-     * 存储过程调用
-     * @access public
-     * @param bool $procedure 是否为存储过程查询
-     * @return $this
-     */
-    public function procedure(bool $procedure = true)
-    {
-        $this->options['procedure'] = $procedure;
-        return $this;
-    }
-
-    /**
      * 指定排序 order('id','desc') 或者 order(['id'=>'desc','create_time'=>'desc'])
      * @access public
      * @param string|array|Raw $field 排序字段
@@ -959,30 +884,6 @@ class BaseQuery
     }
 
     /**
-     * 指定group查询
-     * @access public
-     * @param string|array $group GROUP
-     * @return $this
-     */
-    public function group($group)
-    {
-        $this->options['group'] = $group;
-        return $this;
-    }
-
-    /**
-     * 指定having查询
-     * @access public
-     * @param string $having having
-     * @return $this
-     */
-    public function having(string $having)
-    {
-        $this->options['having'] = $having;
-        return $this;
-    }
-
-    /**
      * 指定查询lock
      * @access public
      * @param bool|string $lock 是否lock
@@ -1000,18 +901,6 @@ class BaseQuery
     }
 
     /**
-     * 指定distinct查询
-     * @access public
-     * @param bool $distinct 是否唯一
-     * @return $this
-     */
-    public function distinct(bool $distinct = true)
-    {
-        $this->options['distinct'] = $distinct;
-        return $this;
-    }
-
-    /**
      * 指定数据表别名
      * @access public
      * @param array|string $alias 数据表别名
@@ -1025,47 +914,6 @@ class BaseQuery
             $table = $this->getTable();
 
             $this->options['alias'][$table] = $alias;
-        }
-
-        return $this;
-    }
-
-    /**
-     * 指定强制索引
-     * @access public
-     * @param string $force 索引名称
-     * @return $this
-     */
-    public function force(string $force)
-    {
-        $this->options['force'] = $force;
-        return $this;
-    }
-
-    /**
-     * 查询注释
-     * @access public
-     * @param string $comment 注释
-     * @return $this
-     */
-    public function comment(string $comment)
-    {
-        $this->options['comment'] = $comment;
-        return $this;
-    }
-
-    /**
-     * 获取执行的SQL语句而不进行实际的查询
-     * @access public
-     * @param bool $fetch 是否返回sql
-     * @return $this|Fetch
-     */
-    public function fetchSql(bool $fetch = true)
-    {
-        $this->options['fetch_sql'] = $fetch;
-
-        if ($fetch) {
-            return new Fetch($this);
         }
 
         return $this;
@@ -1096,66 +944,6 @@ class BaseQuery
     }
 
     /**
-     * 设置自增序列名
-     * @access public
-     * @param string $sequence 自增序列名
-     * @return $this
-     */
-    public function sequence(string $sequence = null)
-    {
-        $this->options['sequence'] = $sequence;
-        return $this;
-    }
-
-    /**
-     * 设置是否REPLACE
-     * @access public
-     * @param bool $replace 是否使用REPLACE写入数据
-     * @return $this
-     */
-    public function replace(bool $replace = true)
-    {
-        $this->options['replace'] = $replace;
-        return $this;
-    }
-
-    /**
-     * 设置当前查询所在的分区
-     * @access public
-     * @param string|array $partition 分区名称
-     * @return $this
-     */
-    public function partition($partition)
-    {
-        $this->options['partition'] = $partition;
-        return $this;
-    }
-
-    /**
-     * 设置DUPLICATE
-     * @access public
-     * @param array|string|Raw $duplicate DUPLICATE信息
-     * @return $this
-     */
-    public function duplicate($duplicate)
-    {
-        $this->options['duplicate'] = $duplicate;
-        return $this;
-    }
-
-    /**
-     * 设置查询的额外参数
-     * @access public
-     * @param string $extra 额外信息
-     * @return $this
-     */
-    public function extra(string $extra)
-    {
-        $this->options['extra'] = $extra;
-        return $this;
-    }
-
-    /**
      * 设置JSON字段信息
      * @access public
      * @param array $json  JSON字段
@@ -1179,46 +967,6 @@ class BaseQuery
     {
         $this->pk = $pk;
         return $this;
-    }
-
-    /**
-     * 指定数据表自增主键
-     * @access public
-     * @param string $autoinc 自增键
-     * @return $this
-     */
-    public function autoinc(string $autoinc)
-    {
-        $this->autoinc = $autoinc;
-        return $this;
-    }
-
-    /**
-     * 获取当前数据表的主键
-     * @access public
-     * @return string|array
-     */
-    public function getPk()
-    {
-        if (empty($this->pk)) {
-            $this->pk = $this->connection->getPk($this->getTable());
-        }
-
-        return $this->pk;
-    }
-
-    /**
-     * 获取当前数据表的自增主键
-     * @access public
-     * @return string
-     */
-    public function getAutoInc()
-    {
-        if (empty($this->autoinc)) {
-            $this->autoinc = $this->connection->getAutoInc($this->getTable());
-        }
-
-        return $this->autoinc;
     }
 
     /**
@@ -1246,17 +994,6 @@ class BaseQuery
         }
 
         return $this->options[$name] ?? null;
-    }
-
-    /**
-     * 获取当前的查询标识
-     * @access public
-     * @param mixed $data 要序列化的数据
-     * @return string
-     */
-    public function getQueryGuid($data = null): string
-    {
-        return md5($this->getConfig('database') . serialize(var_export($data ?: $this->options, true)) . serialize($this->getBind(false)));
     }
 
     /**
@@ -1567,18 +1304,6 @@ class BaseQuery
         }
 
         return true;
-    }
-
-    /**
-     * 创建子查询SQL
-     * @access public
-     * @param bool $sub 是否添加括号
-     * @return string
-     * @throws Exception
-     */
-    public function buildSql(bool $sub = true): string
-    {
-        return $sub ? '( ' . $this->fetchSql()->select() . ' )' : $this->fetchSql()->select();
     }
 
     /**
