+++ src/db/Connection.php
@@ -15,10 +15,6 @@ namespace think\db;
 use Psr\SimpleCache\CacheInterface;
 use think\DbManager;
 use think\db\CacheItem;
-use think\db\exception\DataNotFoundException;
-use think\db\exception\DbException as Exception;
-use think\db\exception\ModelNotFoundException;
-use think\db\exception\PDOException;
 
 /**
  * 数据库连接基础类
@@ -117,49 +113,6 @@ abstract class Connection
     protected $cache;
 
     /**
-     * 架构函数 读取数据库配置信息
-     * @access public
-     * @param array $config 数据库配置数组
-     */
-    public function __construct(array $config = [])
-    {
-        if (!empty($config)) {
-            $this->config = array_merge($this->config, $config);
-        }
-
-        // 创建Builder对象
-        $class = $this->getBuilderClass();
-
-        $this->builder = new $class($this);
-
-        // 执行初始化操作
-        $this->initialize();
-    }
-
-    /**
-     * 初始化
-     * @access protected
-     * @return void
-     */
-    protected function initialize()
-    {
-    }
-
-    /**
-     * 获取当前连接器类对应的Query类
-     * @access public
-     * @return string
-     */
-    abstract public function getQueryClass(): string;
-
-    /**
-     * 获取当前连接器类对应的Builder类
-     * @access public
-     * @return string
-     */
-    abstract public function getBuilderClass(): string;
-
-    /**
      * 获取当前的builder实例对象
      * @access public
      * @return Builder
@@ -226,180 +179,6 @@ abstract class Connection
     {
         $this->config = array_merge($this->config, $config);
     }
-
-    /**
-     * 连接数据库方法
-     * @access public
-     * @param array   $config  接参数
-     * @param integer $linkNum 连接序号
-     * @return mixed
-     * @throws Exception
-     */
-    abstract public function connect(array $config = [], $linkNum = 0);
-
-    /**
-     * 释放查询结果
-     * @access public
-     */
-    abstract public function free();
-
-    /**
-     * 查找单条记录
-     * @access public
-     * @param BaseQuery $query 查询对象
-     * @return array
-     * @throws DbException
-     * @throws ModelNotFoundException
-     * @throws DataNotFoundException
-     */
-    abstract public function find(BaseQuery $query): array;
-
-    /**
-     * 使用游标查询记录
-     * @access public
-     * @param BaseQuery $query 查询对象
-     * @return \Generator
-     */
-    abstract public function cursor(BaseQuery $query);
-
-    /**
-     * 查找记录
-     * @access public
-     * @param BaseQuery $query 查询对象
-     * @return array
-     * @throws DbException
-     * @throws ModelNotFoundException
-     * @throws DataNotFoundException
-     */
-    abstract public function select(BaseQuery $query): array;
-
-    /**
-     * 插入记录
-     * @access public
-     * @param BaseQuery   $query        查询对象
-     * @param boolean $getLastInsID 返回自增主键
-     * @return mixed
-     */
-    abstract public function insert(BaseQuery $query, bool $getLastInsID = false);
-
-    /**
-     * 批量插入记录
-     * @access public
-     * @param BaseQuery   $query   查询对象
-     * @param mixed   $dataSet 数据集
-     * @return integer
-     * @throws \Exception
-     * @throws \Throwable
-     */
-    abstract public function insertAll(BaseQuery $query, array $dataSet = []): int;
-
-    /**
-     * 更新记录
-     * @access public
-     * @param BaseQuery $query 查询对象
-     * @return integer
-     * @throws Exception
-     * @throws PDOException
-     */
-    abstract public function update(BaseQuery $query): int;
-
-    /**
-     * 删除记录
-     * @access public
-     * @param BaseQuery $query 查询对象
-     * @return int
-     * @throws Exception
-     * @throws PDOException
-     */
-    abstract public function delete(BaseQuery $query): int;
-
-    /**
-     * 得到某个字段的值
-     * @access public
-     * @param BaseQuery  $query   查询对象
-     * @param string $field   字段名
-     * @param mixed  $default 默认值
-     * @param bool   $one     返回一个值
-     * @return mixed
-     */
-    abstract public function value(BaseQuery $query, string $field, $default = null);
-
-    /**
-     * 得到某个列的数组
-     * @access public
-     * @param BaseQuery  $query  查询对象
-     * @param string $column 字段名 多个字段用逗号分隔
-     * @param string $key    索引
-     * @return array
-     */
-    abstract public function column(BaseQuery $query, string $column, string $key = ''): array;
-
-    /**
-     * 执行数据库事务
-     * @access public
-     * @param callable $callback 数据操作方法回调
-     * @return mixed
-     * @throws PDOException
-     * @throws \Exception
-     * @throws \Throwable
-     */
-    abstract public function transaction(callable $callback);
-
-    /**
-     * 启动事务
-     * @access public
-     * @return void
-     * @throws \PDOException
-     * @throws \Exception
-     */
-    abstract public function startTrans();
-
-    /**
-     * 用于非自动提交状态下面的查询提交
-     * @access public
-     * @return void
-     * @throws PDOException
-     */
-    abstract public function commit();
-
-    /**
-     * 事务回滚
-     * @access public
-     * @return void
-     * @throws PDOException
-     */
-    abstract public function rollback();
-
-    /**
-     * 关闭数据库（或者重新连接）
-     * @access public
-     * @return $this
-     */
-    abstract public function close();
-
-    /**
-     * 获取最近一次查询的sql语句
-     * @access public
-     * @return string
-     */
-    abstract public function getLastSql(): string;
-
-    /**
-     * 获取最近插入的ID
-     * @access public
-     * @param BaseQuery  $query  查询对象
-     * @param string $sequence 自增序列名
-     * @return mixed
-     */
-    abstract public function getLastInsID(BaseQuery $query, string $sequence = null);
-
-    /**
-     * 初始化数据库连接
-     * @access protected
-     * @param boolean $master 是否主服务器
-     * @return void
-     */
-    abstract protected function initConnect(bool $master = true);
 
     /**
      * 数据库SQL监控
@@ -520,40 +299,14 @@ abstract class Connection
     }
 
     /**
-     * 启动XA事务
+     * 获取返回或者影响的记录数
      * @access public
-     * @param  string $xid XA事务id
-     * @return void
+     * @return integer
      */
-    public function startTransXa(string $xid)
-    {}
-
-    /**
-     * 预编译XA事务
-     * @access public
-     * @param  string $xid XA事务id
-     * @return void
-     */
-    public function prepareXa(string $xid)
-    {}
-
-    /**
-     * 提交XA事务
-     * @access public
-     * @param  string $xid XA事务id
-     * @return void
-     */
-    public function commitXa(string $xid)
-    {}
-
-    /**
-     * 回滚XA事务
-     * @access public
-     * @param  string $xid XA事务id
-     * @return void
-     */
-    public function rollbackXa(string $xid)
-    {}
+    public function getNumRows(): int
+    {
+        return $this->numRows;
+    }
 
     /**
      * 析构方法
