+++ src/db/concern/ParamsBind.php
@@ -47,7 +47,7 @@ trait ParamsBind
      */
     public function bindValue($value, int $type = null, string $name = null)
     {
-        $name = $name ?: 'ThinkBind_' . (count($this->bind) + 1) . '_';
+        $name = $name ?: 'ThinkBind_' . (count($this->bind) + 1) . '_' . mt_rand() . '_';
 
         $this->bind[$name] = [$value, $type ?: PDO::PARAM_STR];
         return $name;
